增加了CRC32校验功能，优化了表解析方法
1. 修改了各个结构体中的CRC32字段类型，从unsigned int改为unsigned long，使其在各个环境下都有至少32位长度；
2. 修改了CRC32的输出语句（格式以及宽度）；
3. 增加了DVB-CRC_32校验文件、方法； 
4. 修改了GetOneSection文件中的部分函数名，使其更准确；
5. 修改了部分注释；
6. 将解析描述流的方法分解；
7. 补全了部分 if的else结构

+++ Parse_EIT.h
@@ -42,7 +42,7 @@ typedef struct TS_EIT_T
 	unsigned int uiSegment_last_section_number :8;
 	unsigned int uiLast_table_id :8;
 	EIT_INFO_T astEIT_info[EIT_INFO_MAX];
-	unsigned int uiCRC_32 :32;
+	unsigned long uiCRC_32 :32;
 } TS_EIT_T;
 
 int ParseEIT_Section(TS_EIT_T *pstTS_EIT, unsigned char *pucSectionBuffer);
